[testcase]: Refactor MACsec test (#5271)

What is the motivation for this PR?
Refactor MACsec test to abstract some common functions from MACsec test.

How did you do it?
Abstract some common functions from test_macsec.py to macsec_*_helper.py

How did you verify/test it?
Check Azp status

Any platform specific information?
Supported testbed topology if it's a new test case?
Documentation
